Refactor simulation using rust modules
